requirements.txt
python-dotenv==1.0.0
slack-bolt==1.18.0
tinydb==4.8.0